Moved most PerformanceCounter logic out of SCore into the new PerformanceCounterManager, some namespace refactoring
