+++ src/SMAPI/Framework/PerformanceCounter/EventPerformanceCounterCategory.cs
