+++ src/SMAPI/Framework/PerformanceCounter/PerformanceCounterEntry.cs
